fix 相关搜索 caused 500 error in filter_scholar
for some search key words,such as fae,due to their small number of results,
google will give a section named '相关搜索',which has the same class with
paper's block but doesn't have a h3 child div.
Because the code didn't check whether a h3 element is exsited, when this
happend, The code will directed visit the h3.a,which can cause a 500 erro.

This fix add h3 exsitence check,fix this issue.